Moved all the unused parts of code to FutureImprovements. Added a ProjectSetup file with the proper prerequisites.
