+++ ProjectSetup.txt
@@ -1,0 +1,13 @@
+VS 2022 Community
+VS Code
+SSMS 2019
+
+nvm - 1.1.10
+node - v12.22.12
+npm - 6.14.16
+
+nvm use 12.22.12
+npm install - inside front project
+
+Configure user secrets connection string
+Authorize IP address on Azure
